+++ package.json
@@ -17,10 +17,12 @@
     "normalize.css": "^8.0.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
+    "react-contenteditable": "^3.3.7",
     "react-cookie": "^4.1.1",
     "react-countdown": "^2.3.5",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
+    "react-quill": "^2.0.0",
     "react-router-dom": "^6.11.0",
     "react-scripts": "^5.0.1",
     "react-slugify": "^3.0.2",
