+++ src/components/index.js
@@ -8,6 +8,3 @@ export { default as QuestionGroup } from "./QuestionGroup";
 export { default as CountDownTimer } from "./CountDownTimer";
 export { default as ListPart } from "./ListPart";
 export { default as MyTable } from "./MyTable";
-export { default as UploadTest } from "./UploadTest";
-export { default as UploadBlog } from "./UploadBlog";
-
