+++ src/pages/Blog/Blog.module.scss
@@ -28,11 +28,15 @@
         }
     }
 
+    .divider {
+        border: 1px solid black;
+        margin-bottom: 1rem;
+    }
 
     .list_blog {
         width: 40%;
         margin-top: 2rem;
-       
+
         .blog {
             border-radius: 10px;
             margin-bottom: 3rem;
