+++ src/pages/ManageBlog/ManageBlog.module.scss
@@ -1,6 +1,6 @@
 .app {
     // display: flex;
-    margin-top: 10%;
+    margin-top: 5%;
     margin-left: 3rem;
 }
