+++ src/pages/ManageTest/ManageTest.module.scss
@@ -1,0 +1,49 @@
+.app {
+    // display: flex;
+    margin-top: 5%;
+    margin-left: 3rem;
+}
+
+.container {
+    width: 100%;
+    margin-left: 8%;
+
+    .taskbar {
+        margin-bottom: 1rem;
+
+        div {
+            display: inline-block;
+            margin-right: 1rem;
+
+
+            .search_bar {
+                display: flex;
+                align-items: center;
+                border: 1px solid black;
+                border-radius: 10px;
+                padding: 0.5rem 1rem;
+            }
+
+            .search_bar input[type="text"] {
+                flex: 1;
+                border: none;
+                outline: none;
+            }
+
+            .search_bar button {
+                border: none;
+                background-color: transparent;
+                cursor: pointer;
+                margin-left: 5px;
+            }
+
+           
+        }
+
+        .filter {
+            border: 1px black solid;
+            padding: 0.8rem 1rem;
+            border-radius: 10px;
+        }
+    }
+}
