+++ src/pages/UploadBlog/UploadBlog.module.scss
@@ -1,3 +1,14 @@
+.app {
+    // display: flex;
+    margin-top: 5%;
+    margin-left: 3rem;
+}
+
+.container {
+    width: 100%;
+    margin-left: 8%;
+}
+
 form {
     margin-top: 20px;
 }
