+++ src/pages/UploadTest/UploadTest.module.scss
@@ -1,3 +1,14 @@
+.app {
+    // display: flex;
+    margin-top: 5%;
+    margin-left: 3rem;
+}
+
+.container {
+    width: 100%;
+    margin-left: 8%;
+}
+
 .upload_form {
     display: flex;
     flex-direction: column;
@@ -15,6 +26,7 @@
 
     div {
         overflow: hidden;
+
         p {
             font-size: 2rem;
             text-align: center;
@@ -49,6 +61,14 @@ input[type="file"] {
     background-color: #005ea3;
 }
 
+.labeltext {
+    margin-top: 5px;
+    font-weight: bold;
+}
+
+.labeltext+input {
+    height: 50px;
+}
 
 .inputfile {
     width: 0.1px;
